backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database (SQLModel + async PostgreSQL driver)
sqlmodel>=0.0.16
sqlalchemy>=2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI Integration
langchain
langchain-openai
langchain-google-genai
google-generativeai>=0.3.0
openai

# Validation
pydantic>=2.5.3
pydantic-settings>=2.1.0
email-validator==2.1.0

# Utils
python-dotenv==1.0.0
httpx==0.26.0
redis==5.0.1
